Clean up implementation of serial communication.
The section of code now will actually terminate and not loop forever.
The legacy chunk of code which was commented out has been removed.

Current thoughts are how to figure out how many characters are
available for read -- the UP core seems to have it, but I believe
the register map is actually different. Would consider exploring the
end of packet character option tomorrow.

Current implementation has it reading one character at a time and
adding that character to an array-this means in our timer interrupt
our context actually needs something passed into it-our counter.

Our timer ISR should return when a special character is found, set
some flag and let the main function do something about it. Like
print something to the LCD, or something.