Add basic VGA functionality. VGA functions being implemented. Serial only reads 63 characters at a time. Button interrupt fixed.
There seems to be a character limit of 63 on the rs232 FIFO. I
think there's a chance that the issue can be resolved if we use the
Interrupts from the rs232 to read whenever available, and then look
for an end character (whether in hardware or whatever). Attempts to
use a loop structure to grab more values is thwarted by the fact
that characters are lost if the buffer is full.

VGA basics are already there. Might draw some boxes, but will have
to see how much time we have. I'd rather try to add audio functions
as an alarm when the medication happens.

There seems to have been an issue in the way we handled button
interrupts where we didn't clear properly, allowing the IRQ to be
HIGH even though we disabled the IRQ. When the IRQ was re-enabled,
the push button ISR would trigger immediately. This appears to be
fixed. Let's hope it stays that way.